Working with negative angles

+++ apps/abs_controller/abs_manager.h
@@ -15,9 +15,9 @@
  * @param carVel            Car velocity (max of wheel signals)
  */
 typedef struct {
-    uint32_t wheelVel[NUM_WHEELS];
-    uint32_t steeringAngle;
-    uint32_t carVel;
+    int32_t wheelVel[NUM_WHEELS];
+    float steeringAngle;
+    int32_t carVel;
     bool sold;
 } CarAttributes_t;
